refactor: move helpers to _helpers/, rename config_key to settings_key, update docs and tests
