+++ packages/kiarina-lib-google-cloud-storage/src/kiarina/lib/google/cloud_storage/_helpers/get_storage_client.py
@@ -5,8 +5,8 @@ from kiarina.lib.google.auth import get_credentials
 
 
 def get_storage_client(
-    auth_config_key: str | None = None,
+    auth_settings_key: str | None = None,
     **kwargs: Any,
 ) -> storage.Client:
-    credentials = get_credentials(auth_config_key)
+    credentials = get_credentials(auth_settings_key)
     return storage.Client(credentials=credentials, **kwargs)
